removing setra css file in public folder that wasnt in css folder
