--- public/style.css
@@ -1,3 +1,0 @@
-body {
-    color:#D6D6D6;
-}
