+++ server.js
@@ -1,26 +1,26 @@
-// Dependencies=============================================================
+const path = require('path');
 const express = require('express');
-const path = require("path");
+const exphbs = require('express-handlebars');
+const routes = require('./controllers');
+const helpers = require('./utils/helpers');
+const sequelize = require('./config/connection');
 
-// Sets up the Express App
-// =============================================================
 const app = express();
-const port = process.env.port || 3000;
+const PORT = process.env.PORT || 3000;
 
-const handlebars = require('express-handlebars');
+// Create the Handlebars.js engine object with custom helper functions
+const hbs = exphbs.create({ helpers });
 
+// Inform Express.js which template engine we're using
+app.engine('handlebars', hbs.engine);
 app.set('view engine', 'handlebars');
 
-app.engine('handlebars', handlebars({
-    layoutsDir: '${__dirname}/views/layouts'
-}));
+app.use(express.json());
+app.use(express.urlencoded({ extended: true }));
+app.use(express.static(path.join(__dirname, 'public')));
 
-app.use(express.static('public'));
+app.use(routes);
 
-app.get('/', (req, res) => {
-    res.render('main', {layout: 'index'});
-});
-
-app.listen(port, () => {
-    console.log('app listening to port ' + port);
+sequelize.sync({ force: false }).then(() => {
+  app.listen(PORT, () => console.log('Now listening to port: ' + PORT));
 });
